+++ packages/google-ai-generativelanguage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.3.1](https://github.com/googleapis/google-cloud-python/compare/google-ai-generativelanguage-v0.3.0...google-ai-generativelanguage-v0.3.1) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11444](https://github.com/googleapis/google-cloud-python/issues/11444)) ([9aa301a](https://github.com/googleapis/google-cloud-python/commit/9aa301ae6ca3080cae286a19de9cdc1b796ab37d))
+
 ## [0.3.0](https://github.com/googleapis/google-cloud-python/compare/google-ai-generativelanguage-v0.2.1...google-ai-generativelanguage-v0.3.0) (2023-06-29)