+++ packages/google-cloud-api-keys/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.5.3](https://github.com/googleapis/google-cloud-python/compare/google-cloud-api-keys-v0.5.2...google-cloud-api-keys-v0.5.3) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11444](https://github.com/googleapis/google-cloud-python/issues/11444)) ([9aa301a](https://github.com/googleapis/google-cloud-python/commit/9aa301ae6ca3080cae286a19de9cdc1b796ab37d))
+
 ## [0.5.2](https://github.com/googleapis/python-api-keys/compare/v0.5.1...v0.5.2) (2023-03-23)