+++ packages/google-cloud-billing/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.11.1](https://github.com/googleapis/google-cloud-python/compare/google-cloud-billing-v1.11.0...google-cloud-billing-v1.11.1) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11445](https://github.com/googleapis/google-cloud-python/issues/11445)) ([98bddda](https://github.com/googleapis/google-cloud-python/commit/98bdddafc821e2fc6e86a31965da0c46899aa229))
+
 ## [1.11.0](https://github.com/googleapis/google-cloud-python/compare/google-cloud-billing-v1.10.1...google-cloud-billing-v1.11.0) (2023-06-19)