+++ packages/google-cloud-certificate-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.4.3](https://github.com/googleapis/google-cloud-python/compare/google-cloud-certificate-manager-v1.4.2...google-cloud-certificate-manager-v1.4.3) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11445](https://github.com/googleapis/google-cloud-python/issues/11445)) ([98bddda](https://github.com/googleapis/google-cloud-python/commit/98bdddafc821e2fc6e86a31965da0c46899aa229))
+
 ## [1.4.2](https://github.com/googleapis/python-certificate-manager/compare/v1.4.1...v1.4.2) (2023-03-23)