+++ packages/google-cloud-data-fusion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.8.2](https://github.com/googleapis/google-cloud-python/compare/google-cloud-data-fusion-v1.8.1...google-cloud-data-fusion-v1.8.2) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11445](https://github.com/googleapis/google-cloud-python/issues/11445)) ([98bddda](https://github.com/googleapis/google-cloud-python/commit/98bdddafc821e2fc6e86a31965da0c46899aa229))
+
 ## [1.8.1](https://github.com/googleapis/python-data-fusion/compare/v1.8.0...v1.8.1) (2023-03-23)