+++ packages/google-cloud-resource-manager/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 [1]: https://pypi.org/project/google-cloud-resource-manager/#history
 
+## [1.10.2](https://github.com/googleapis/google-cloud-python/compare/google-cloud-resource-manager-v1.10.1...google-cloud-resource-manager-v1.10.2) (2023-07-05)
+
+
+### Bug Fixes
+
+* Add async context manager return types ([#11448](https://github.com/googleapis/google-cloud-python/issues/11448)) ([4d2c35a](https://github.com/googleapis/google-cloud-python/commit/4d2c35a1cd0b68b6d481d5611ff820451273e859))
+
 ## [1.10.1](https://github.com/googleapis/python-resource-manager/compare/v1.10.0...v1.10.1) (2023-05-04)